Add padic_types.h and reduce header dependencies from padic modules
