+++ src/padic_mat/add.c
@@ -9,8 +9,10 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz.h"
 #include "fmpz_vec.h"
 #include "fmpz_mat.h"
+#include "padic.h"
 #include "padic_mat.h"
 
 /*
