+++ src/padic_mat/canonicalise.c
@@ -9,6 +9,7 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
 #include "fmpz_mat.h"
 #include "padic_poly.h"
 #include "padic_mat.h"
