+++ src/padic_mat/get_entry_padic.c
@@ -9,7 +9,8 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
-#include "fmpz_mat.h"
+#include "fmpz.h"
+#include "padic.h"
 #include "padic_mat.h"
 
 void padic_mat_get_entry_padic(padic_t rop,
