+++ src/padic_mat/set_fmpq_mat.c
@@ -11,6 +11,7 @@
 
 #include "fmpq.h"
 #include "fmpq_mat.h"
+#include "padic.h"
 #include "padic_mat.h"
 
 void padic_mat_set_fmpq_mat(padic_mat_t B,
