+++ src/padic_poly/evaluate_padic.c
@@ -9,7 +9,9 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
 #include "fmpz_mod_poly.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 static void _fmpz_mod_poly_evaluate_fmpz_horner(fmpz_t res, const fmpz *poly, slong len,
