+++ src/padic_poly/get_fmpq_poly.c
@@ -9,6 +9,7 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
 #include "fmpq_poly.h"
 #include "padic_poly.h"
