+++ src/padic_poly/inv_series.c
@@ -10,7 +10,9 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
 #include "fmpz_poly.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 /*  The tests fail if one reduces the coefficients modulo N before
