+++ src/padic_poly/io.c
@@ -10,6 +10,8 @@
 */
 
 #include <stdio.h>
+#include "fmpz_vec.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 int _padic_poly_fprint(FILE *file, const fmpz *poly, slong val, slong len,
@@ -194,3 +196,22 @@ int _padic_poly_print(const fmpz *poly, slong val, slong len, const padic_ctx_t 
 int padic_poly_print(const padic_poly_t poly, const padic_ctx_t ctx) { return padic_poly_fprint(stdout, poly, ctx); }
 int _padic_poly_print_pretty(const fmpz *poly, slong val, slong len, const char *var, const padic_ctx_t ctx) { return _padic_poly_fprint_pretty(stdout, poly, val, len, var, ctx); }
 int padic_poly_print_pretty(const padic_poly_t poly, const char *var, const padic_ctx_t ctx) { return padic_poly_fprint_pretty(stdout, poly, var, ctx); }
+
+int padic_poly_debug(const padic_poly_t poly)
+{
+    flint_printf("(alloc = %wd, length = %wd, val = %wd, N = %wd, vec = ",
+        poly->alloc, poly->length, poly->val, poly->N);
+    if (poly->coeffs)
+    {
+        flint_printf("{");
+        _fmpz_vec_print(poly->coeffs, poly->alloc);
+        flint_printf("}");
+    }
+    else
+    {
+        flint_printf("NULL");
+    }
+    flint_printf(")");
+
+    return 1;
+}
