+++ src/padic_poly/is_reduced.c
@@ -9,6 +9,7 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "padic.h"
 #include "padic_poly.h"
 
 int _padic_poly_is_reduced(const fmpz *op, slong val, slong len, slong N,
