+++ src/padic_poly/reduce.c
@@ -9,6 +9,8 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 void padic_poly_reduce(padic_poly_t poly, const padic_ctx_t ctx)
