+++ src/padic_poly/scalar_mul_padic.c
@@ -9,6 +9,8 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 void _padic_poly_scalar_mul_padic(fmpz *rop, slong *rval, slong N,
