+++ src/padic_poly/set.c
@@ -9,6 +9,8 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 void padic_poly_set(padic_poly_t poly1,
