+++ src/padic_poly/set_coeff_padic.c
@@ -9,6 +9,8 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "fmpz_vec.h"
+#include "padic.h"
 #include "padic_poly.h"
 
 void padic_poly_set_coeff_padic(padic_poly_t poly, slong n, const padic_t x,
