+++ src/padic_poly/set_fmpz.c
@@ -9,6 +9,7 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include "padic.h"
 #include "padic_poly.h"
 
 void padic_poly_set_fmpz(padic_poly_t poly, const fmpz_t x,
